+++ classes/booking_settings.php
@@ -371,6 +371,9 @@ class booking_settings {
     /** @var string $customfieldsforfilter */
     public $customfieldsforfilter = null;
 
+    /** @var array $fulltextsearchcolumns columns added to the full text search of the booking options table */
+    public $fulltextsearchcolumns = null;
+
     /** @var string $toporientation */
     public $toporientation = null;